Fix audits index - Task 35436

+++ spec/dummy/db/migrate/20211202111615_fix_audits_index.rb
@@ -1,6 +1,9 @@
 class FixAuditsIndex < ActiveRecord::Migration[6.1]
   def change
-    remove_index :audits, name: 'index_audits_on_version'
+    if index_exists?('audits', 'version')
+      remove_index :audits, name: 'index_audits_on_version'
+    end
+
     remove_index :audits, name: 'auditable_index'
 
     add_index :audits, %i[auditable_id auditable_type version], name: 'auditable_index'
